Use JUnit 5 for testing
* Also move restapi tests to their proper location in the tree

* Run Integration Tests in `testing` in one JVM to save on
  restarting the backend cluster.

* Also run `testing` suites in parallel. Being backbox tests
  they should be able to run concurrently.

* Note: testcontainers are removed because they transitively
  pull in JUnit 4 (they are no longer used anyway).
