+++ cql/src/test/java/org/apache/cassandra/stargate/transport/internal/PlainTextTokenSaslNegotiatorTest.java
@@ -18,7 +18,7 @@ import java.nio.charset.StandardCharsets;
 import org.apache.cassandra.stargate.exceptions.AuthenticationException;
 import org.apache.cassandra.stargate.transport.internal.PlainTextTokenSaslNegotiator.Credentials;
 import org.apache.commons.lang3.StringUtils;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class PlainTextTokenSaslNegotiatorTest {
   private final String TOKEN = "a24b121a-a385-44a6-8ae1-fe7542dbc490";
