+++ persistence-cassandra-3.11/pom.xml
@@ -96,12 +96,6 @@
       <version>${junit.version}</version>
       <scope>test</scope>
     </dependency>
-    <dependency>
-      <groupId>org.testcontainers</groupId>
-      <artifactId>testcontainers</artifactId>
-      <version>1.14.3</version>
-      <scope>test</scope>
-    </dependency>
   </dependencies>
   <build>
     <plugins>
