+++ testing/src/test/java/io/stargate/it/PersistenceTest.java
@@ -24,12 +24,32 @@ import static io.stargate.db.datastore.schema.Column.Kind.PartitionKey;
 import static io.stargate.db.datastore.schema.Column.Kind.Static;
 import static io.stargate.db.datastore.schema.Column.Order.Asc;
 import static io.stargate.db.datastore.schema.Column.Order.Desc;
-import static io.stargate.db.datastore.schema.Column.Type.*;
+import static io.stargate.db.datastore.schema.Column.Type.Ascii;
+import static io.stargate.db.datastore.schema.Column.Type.Bigint;
+import static io.stargate.db.datastore.schema.Column.Type.Blob;
 import static io.stargate.db.datastore.schema.Column.Type.Boolean;
+import static io.stargate.db.datastore.schema.Column.Type.Date;
+import static io.stargate.db.datastore.schema.Column.Type.Decimal;
 import static io.stargate.db.datastore.schema.Column.Type.Double;
+import static io.stargate.db.datastore.schema.Column.Type.Duration;
 import static io.stargate.db.datastore.schema.Column.Type.Float;
+import static io.stargate.db.datastore.schema.Column.Type.Inet;
+import static io.stargate.db.datastore.schema.Column.Type.Int;
+import static io.stargate.db.datastore.schema.Column.Type.List;
+import static io.stargate.db.datastore.schema.Column.Type.Map;
+import static io.stargate.db.datastore.schema.Column.Type.Set;
+import static io.stargate.db.datastore.schema.Column.Type.Smallint;
+import static io.stargate.db.datastore.schema.Column.Type.Text;
+import static io.stargate.db.datastore.schema.Column.Type.Time;
+import static io.stargate.db.datastore.schema.Column.Type.Timestamp;
+import static io.stargate.db.datastore.schema.Column.Type.Timeuuid;
+import static io.stargate.db.datastore.schema.Column.Type.Tinyint;
+import static io.stargate.db.datastore.schema.Column.Type.Tuple;
+import static io.stargate.db.datastore.schema.Column.Type.Uuid;
+import static io.stargate.db.datastore.schema.Column.Type.Varchar;
+import static io.stargate.db.datastore.schema.Column.Type.Varint;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.fail;
+import static org.assertj.core.api.Fail.fail;
 
 import com.datastax.oss.driver.api.core.data.CqlDuration;
 import com.datastax.oss.driver.api.core.data.TupleValue;
@@ -53,6 +73,7 @@ import io.stargate.db.datastore.schema.ParameterizedType;
 import io.stargate.db.datastore.schema.Schema;
 import io.stargate.db.datastore.schema.Table;
 import io.stargate.db.datastore.schema.UserDefinedType;
+import java.lang.reflect.Method;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.net.Inet4Address;
@@ -68,6 +89,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
@@ -76,11 +98,10 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import net.jcip.annotations.NotThreadSafe;
 import org.javatuples.Pair;
-import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TestName;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 import org.osgi.framework.InvalidSyntaxException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -89,23 +110,23 @@ import org.slf4j.LoggerFactory;
 public class PersistenceTest extends BaseOsgiIntegrationTest {
   private static final Logger logger = LoggerFactory.getLogger(PersistenceTest.class);
 
-  @Rule public TestName name = new TestName();
-
   private DataStore dataStore;
   private String table;
   private String keyspace;
 
   private static final int CUSTOM_PAGE_SIZE = 50;
 
-  @Before
-  public void setup() throws InvalidSyntaxException {
+  @BeforeEach
+  public void setup(TestInfo testInfo) throws InvalidSyntaxException {
     Persistence persistence = getOsgiService("io.stargate.db.Persistence", Persistence.class);
     ClientState clientState = persistence.newClientState("");
     QueryState queryState = persistence.newQueryState(clientState);
     dataStore = persistence.newDataStore(queryState, null);
     logger.info("{} {} {}", clientState, queryState, dataStore);
 
-    String testName = name.getMethodName();
+    Optional<String> name = testInfo.getTestMethod().map(Method::getName);
+    assertThat(name).isPresent();
+    String testName = name.get();
     keyspace = "ks_" + testName;
     table = testName;
   }
@@ -221,7 +242,7 @@ public class PersistenceTest extends BaseOsgiIntegrationTest {
                 .table(table));
   }
 
-  @Ignore("Disabling for now since it currently just hangs")
+  @Disabled("Disabling for now since it currently just hangs")
   @Test
   public void testInsertingAndReadingDifferentTypes() throws Exception {
     Keyspace ks = createKeyspace();
@@ -504,7 +525,7 @@ public class PersistenceTest extends BaseOsgiIntegrationTest {
                 .toString());
   }
 
-  @Ignore("Disabling for now since it currently just hangs")
+  @Disabled("Disabling for now since it currently just hangs")
   @Test
   public void testTupleWithAllSimpleTypes()
       throws UnknownHostException, ExecutionException, InterruptedException {
@@ -552,7 +573,7 @@ public class PersistenceTest extends BaseOsgiIntegrationTest {
     }
   }
 
-  @Ignore("Disabling UDT related tests for now")
+  @Disabled("Disabling UDT related tests for now")
   @Test
   public void testUDTWithAllSimpleTypes()
       throws UnknownHostException, ExecutionException, InterruptedException {
@@ -639,7 +660,7 @@ public class PersistenceTest extends BaseOsgiIntegrationTest {
         .build();
   }
 
-  @Ignore("Disabling for now since it currently just hangs")
+  @Disabled("Disabling for now since it currently just hangs")
   @Test
   public void testUDT() throws ExecutionException, InterruptedException {
     Keyspace ks = createKeyspace();
@@ -697,7 +718,7 @@ public class PersistenceTest extends BaseOsgiIntegrationTest {
     //        assertThat(udtValue.getTupleValue("mytuple")).isEqualTo(tuple);
   }
 
-  @Ignore("Disabling for now since it currently just hangs")
+  @Disabled("Disabling for now since it currently just hangs")
   @Test
   public void testTupleMismatch() throws ExecutionException, InterruptedException {
     Keyspace ks = createKeyspace();
@@ -756,7 +777,7 @@ public class PersistenceTest extends BaseOsgiIntegrationTest {
     }
   }
 
-  @Ignore("Disabling for now since it fails with a strange MV schema generated")
+  @Disabled("Disabling for now since it fails with a strange MV schema generated")
   @Test
   public void testMvIndexes() throws ExecutionException, InterruptedException {
     createKeyspace();
