+++ testing/src/test/java/io/stargate/it/http/GraphqlTest.java
@@ -61,10 +61,8 @@ import net.jcip.annotations.NotThreadSafe;
 import okhttp3.OkHttpClient;
 import org.apache.http.HttpStatus;
 import org.jetbrains.annotations.NotNull;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TestName;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.osgi.framework.InvalidSyntaxException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -73,15 +71,13 @@ import org.slf4j.LoggerFactory;
 public class GraphqlTest extends BaseOsgiIntegrationTest {
   private static final Logger logger = LoggerFactory.getLogger(GraphqlTest.class);
 
-  @Rule public TestName name = new TestName();
-
   private DataStore dataStore;
   private String keyspace;
   private static String authToken;
   private static final ObjectMapper objectMapper = new ObjectMapper();
   private static String host = "http://" + stargateHost;
 
-  @Before
+  @BeforeEach
   public void setup()
       throws InvalidSyntaxException, ExecutionException, InterruptedException, IOException {
     keyspace = "betterbotz";
